Move all code to kaijuu/ directory
Created a directory named "kaijuu/", moved all code except for nall and
hiro into it, and edited GNUmakefile to refer to nall and hiro outside
of the directory. This mirrors the directory structure that byuu uses
nowadays with higan and bass, as well as hex_usr's nSide and Mercurial
Magic.